ia2TextMozilla: Use IAccessible2_2::accessibleWithCaret to retrieve the caret directly if it's supported.
Normally, we have to call caretOffset on the focus. If the caret is within a descendant, we have to call caretOffset on each embedded object until we get to the deepest object.
We do this using an in-process helper.
However, IAccessible2_2::accessibleWithCaret provides a way to get the caret object and offset directly in a single call.
This is particularly beneficial for performance in Firefox multi-process, since our in-process helper ends up making cross-process calls in that case.